+++ example/main.dart
@@ -5,14 +5,13 @@ import 'package:image/image.dart';
 import 'package:args/args.dart';
 import 'package:path/path.dart' as path;
 
-void main(List<String> args) {
+void main(List<String> args) async {
   final parser = ArgParser()
-    ..addOption('lut-file', abbr: 'l', allowMultiple: false)
-    ..addOption('in-img', abbr: 'i', allowMultiple: true)
-    ..addOption('out-dir',
-        abbr: 'o', defaultsTo: './out', allowMultiple: false);
+    ..addOption('lut-file', abbr: 'l')
+    ..addOption('in-img', abbr: 'i')
+    ..addOption('out-dir', abbr: 'o', defaultsTo: './out');
 
-  _runStreamed(parser.parse(args));
+  _run(parser.parse(args));
 }
 
 void _run(final ArgResults argResult) {
@@ -24,16 +23,16 @@ void _run(final ArgResults argResult) {
     final lut = LUT.fromString(lutFile.readAsStringSync());
 
     lut.isLoaded.listen((data) async {
-      Image image = decodeImage(imageFile.readAsBytesSync());
+      final image = decodeImage(imageFile.readAsBytesSync());
       final sw = Stopwatch()..start();
-      var interp = InterpolationType.trilinear;
+      final interp = InterpolationType.trilinear;
 
-      var v = lut.applySync(image.getBytes(), interp);
+      final v = lut.applySync(image.getBytes(), interp);
 
       print('lut apply in ${sw.elapsed}');
       sw.stop();
-      var image2 = Image.fromBytes(image.width, image.height, v);
-      var outputFile = new File(
+      final image2 = Image.fromBytes(image.width, image.height, v);
+      final outputFile = new File(
           '${argResult['out-dir']}/${path.basename(imageFile.path)}_${path.basename(lutFile.path)}_$interp.jpg')
         ..writeAsBytesSync(encodeJpg(image2));
       print('output image write to: ${outputFile.path}');
@@ -49,14 +48,13 @@ void _runStreamed(final ArgResults argResult) async {
 
     final lut = LUT.fromString(lutFile.readAsStringSync());
 
-    var isLoaded = await lut.awaitLoading();
+    final isLoaded = await lut.awaitLoading();
     if (isLoaded) {
-      Image image = decodeImage(imageFile.readAsBytesSync());
-      var interp = InterpolationType.trilinear;
+      final image = decodeImage(imageFile.readAsBytesSync());
+      final interp = InterpolationType.trilinear;
       lut.applyAsStream(image.getBytes(), interp).listen((result) {
-        print('#${result.toRadixString(16).padLeft(4, '0')}');
+        //print('#${result.toRadixString(16).padLeft(4, '0')}');
       });
     }
-    ;
   }
 }