+++ lib/dart_lut.dart
@@ -2,4 +2,4 @@ library dart_lut;
 
 export 'src/lut.dart';
 export 'src/table.dart';
-export 'src/rgb.dart';
+export 'src/colour.dart';