+++ lib/src/colour.dart
@@ -1,10 +1,13 @@
-///This class stores RGB values
-class RGB {
+/// This class stores RGB values
+class Colour {
   final double _r;
   final double _g;
   final double _b;
 
-  RGB(double this._r, double this._g, double this._b);
+  const Colour(double this._r, double this._g, double this._b)
+      : assert(_r != null),
+        assert(_g != null),
+        assert(_b != null);
 
   /// The Red value
   double get r => _r;
@@ -31,8 +34,8 @@ class RGB {
 
   @override
   bool operator ==(Object other) {
-    if (other is RGB) {
-      return this._r == other.r && this._g == other.g && this._b == other.g;
+    if (other is Colour) {
+      return this._r == other.r && this._g == other.g && this._b == other.b;
     }
     return false;
   }